Add PilotUpload task (wooga/atlas-build-unity#87)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/87

Description
===========

This is the second patch which will soon replace the old `fastlane`
based task from `net.wooga.build-unity-ios`. The new task type
`PilotUpdate` will soon replace the old `PublishTestFlight`
implementation. This new task has nearly the same interface. It uses
like all oter new types the Provider API to get rid of convention
mapping as much as possible. The implementation uses a lot of the
underlying abstract base class `AbstractFastlaneTask` which was
introduced in the last patch. This makes this change rather small. It
supports logfile creation and the ability to add additional arguments in
case fastlane adds more commandline arguments or we need to use a few of
the ones I did not yet implement because of a good reason to have them
in.

Changes
=======

* ![ADD] `PilotUpload` task type