Fix overload for task action for SighRenewBatch (wooga/atlas-build-unity#172)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/172

Description
===========

I was under the assumption what when one overrides a gradle task
type which contains an `TaskAction` annotated method that
if one annotates another new method that the original action is
no longer executed. That is not the case. So I fixed this here
to override the `exec` method rather than declaring another action
to be executed.

Changes
=======

* ![FIX] overload for task action in `SighRenewBatch` task