Fix sigh renew password set through environment
Description
===========

This is a small patch which only involves to bump
`com.wooga.gradle:gradle-commons` to min version `1.5.1`.
We need the functionality of the `ExecSpec`/`ProcessExecutor` to handle
environment values correctly. The latest version had a bug and was not
pushing the values provided in the environment map to the executable.

I adjusted the tests to actually test the values being passed to the
called process instead of checking the task properties only.

Changes
=======

* ![FIX] sigh renew password set through environment