Improve SighRenew task properties (#7)
Description
===========

The fastlane tool added some new arguments and flags to the
`sigh renew` command. I simply added new properties to map these
to the cli arguments.

I decided to refactor the testsetup slightly. This is the actual bulk
of these changes. I wanted a setup similar to other plugins. I didn't
commit on going over each and every test feature. So some is now in
a mix state. But the end result is better than what was there before.

Changes
=======

* ![IMPROVE] `SighRenew` task properties